Melhorias nos response

+++ src/main/java/br/com/projeto/contratados/domain/service/empresa/SetorCargoService.java
@@ -1,7 +1,9 @@
 package br.com.projeto.contratados.domain.service.empresa;
 
 import br.com.projeto.contratados.config.exception.excecoes.AnuncioVagaNaoEncontradoException;
+import br.com.projeto.contratados.config.exception.excecoes.EmpresaNaoEncontradaException;
 import br.com.projeto.contratados.config.exception.excecoes.SetorCargoNaoEncontradoException;
+import br.com.projeto.contratados.config.security.TokenService;
 import br.com.projeto.contratados.domain.entity.empresa.AnuncioVaga;
 import br.com.projeto.contratados.domain.entity.empresa.SetorCargo;
 import br.com.projeto.contratados.domain.repository.empresa.AnuncioVagaRepository;
@@ -24,13 +26,23 @@ public class SetorCargoService {
     @Autowired
     private AnuncioVagaRepository anuncioVagaRepository;
 
+    @Autowired
+    private TokenService tokenService;
+
+    private Integer getIdEmpresa(){
+        return tokenService.getAuthenticatedEmpresa();
+    }
+
     public SetorCargo cadastrar(SetorCargoRequest form) {
 
-        SetorCargo setorCargo = form.converter();
-
-        Optional<AnuncioVaga> optional = anuncioVagaRepository.findById(setorCargo.getAnuncioVaga().getId());
+        Optional<AnuncioVaga> optional = anuncioVagaRepository.findById(form.getAnuncioVaga().getId());
         if (optional.isEmpty())
             throw new AnuncioVagaNaoEncontradoException("Anúncio vaga não encontrado, não foi possível cadastrar um setor e um cargo");
+
+        if (!optional.get().getEmpresa().getId().equals(getIdEmpresa()))
+            throw new EmpresaNaoEncontradaException("Empresa não encontrada, não foi possível cadastrar um setor e um cargo");
+
+        var setorCargo = form.converter();
 
         return setorCargoRepository.save(setorCargo);
     }
@@ -45,7 +57,10 @@ public class SetorCargoService {
         if (optional.isEmpty())
             throw new SetorCargoNaoEncontradoException("Setor Cargo não encontrado, ele não pode ser atualizado");
 
-        SetorCargo setorCargo = form.converter(id, setorCargoRepository);
+        if (!optional.get().getAnuncioVaga().getEmpresa().getId().equals(getIdEmpresa()))
+            throw new EmpresaNaoEncontradaException("Empresa não encontrada, não foi possível atualizar os setores e os cargos");
+
+        var setorCargo = form.converter(id, setorCargoRepository);
         return setorCargoRepository.save(setorCargo);
     }
 
@@ -54,7 +69,10 @@ public class SetorCargoService {
         if (optional.isEmpty())
             throw new SetorCargoNaoEncontradoException("Setor Cargo não encontrado, ele não pode ser deletado");
 
-        SetorCargo setorCargo = setorCargoRepository.getOne(id);
+        if (!optional.get().getAnuncioVaga().getEmpresa().getId().equals(getIdEmpresa()))
+            throw new EmpresaNaoEncontradaException("Empresa não encontrada, não foi possível deletar o setor e o cargo");
+
+        var setorCargo = setorCargoRepository.getOne(id);
         setorCargoRepository.deleteById(id);
         return setorCargo;
     }
